requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-filter==25.2
django-stubs==5.2.8
djangorestframework-stubs==3.16.6
